requirements.txt
fastapi
uvicorn
langchain
python-dotenv
scrapy
beautifulsoup4
youtube-dl
openai
requests
pydantic 
langchain_community
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
google-cloud-aiplatform
langchain-google-genai
google-api-python-client
isodate